Move internal/platform to daemon/internal/platform
