+++ daemon/info.go
@@ -18,10 +18,10 @@ import (
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/internal/filedescriptors"
 	"github.com/docker/docker/daemon/internal/metrics"
+	"github.com/docker/docker/daemon/internal/platform"
 	"github.com/docker/docker/daemon/logger"
 	"github.com/docker/docker/daemon/pkg/registry"
 	"github.com/docker/docker/dockerversion"
-	"github.com/docker/docker/internal/platform"
 	"github.com/docker/docker/pkg/meminfo"
 	"github.com/docker/docker/pkg/parsers/kernel"
 	"github.com/docker/docker/pkg/parsers/operatingsystem"
