+++ daemon/internal/platform/platform_windows.go
